Test for transaction rollback on procedure with specified schema
Hi,

We are experiencing an issue with rolling back a transaction which executes a stored procedure in a schema other than dbo.  I have created a test that reproduces the issue: even if the schema is dbo, by specifying it explicitly we lose the transaction.

I followed the code through as far as DynamicSchema.TryInvokeMember calling command.Execute without passing any transaction?  That's as far as I got, but I could be going down completely the wrong rabbit hole...  I'd be happy to (try to) help further if you agree this is in need of fixing!

Tim